requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.37.10
botocore==1.37.10
certifi==2025.1.31
click==8.1.8
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jmespath==1.0.1
mcp==1.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.6.6
uvicorn==0.34.0
